+++ src/components/ui/textEditor/sets.js
@@ -1,0 +1,10 @@
+import { BoolEnums } from '../../../enums';
+import { ContentSwSides } from '../contentSw/enums';
+
+export const TextFileViewerSettings = {
+    gamePage: {
+        id: 'text-editor',
+        displayLabel: BoolEnums.bFalse,
+        side: ContentSwSides.right
+    }
+};