+++ src/components/ui/textEditor/theme.js
@@ -1,0 +1,17 @@
+import {  PackThemes } from '../../../theme/enums';
+
+export const TextFileViewerTheme = {
+    [PackThemes.common]: {
+        wrapper: {
+            background: '#dcdcdc'
+        },
+        container: {
+            border: 'black'
+        },
+        loading: {
+            background: 'white',
+            text: 'black',
+            border: 'blue',
+        }
+    }
+};