+++ src/pages/safe/editor/files.js
@@ -1,0 +1,36 @@
+import { LabelIcons } from '../../../components/ui/contentSw/enums';
+
+export const textFiles = [
+    { 
+        id: 'file1', 
+        name: 'file1',
+        content: 'Type your content',
+        rows: 20,
+        cols: 60,
+        icon: LabelIcons.text,
+    },
+    {
+        id: 'file2',
+        name: 'file2',
+        content: 'Type your content',
+        rows: 20,
+        cols: 60,
+        icon: LabelIcons.text,
+    },
+    {
+        id: 'file3',
+        name: 'file3',
+        content: 'Type your content',
+        rows: 20,
+        cols: 60,
+        icon: LabelIcons.text,
+    },
+    {
+        id: 'file4',
+        name: 'file4',
+        content: 'Type your content',
+        rows: 20,
+        cols: 60,
+        icon: LabelIcons.text,
+    }
+];