+++ src/pages/safe/index.js
@@ -1,3 +1,4 @@
 import './safeKing/index';
+import './editor/editorPage';
 import './welcome';
 import './tabs';