+++ src/pages/safe/sets.js
@@ -2,5 +2,6 @@ export const SafePageTabs = {
     tabLinks: {
         game: 'game-btn',
         welcome: 'welcome-btn',
+        editor: 'text-editor-btn'
     }
 };